# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Pinjie Fullstack Blog Backend",
    "description": "独立个人博客的后端 API，当前提供认证、账户、后台管理、文件资产和系统设置基础能力。",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/users/me": {
      "get": {
        "tags": [
          "用户账户"
        ],
        "summary": "获取当前用户",
        "operationId": "get_me_api_v1_users_me_get",
        "parameters": [
          {
            "name": "pinjie_blog_web_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UserPrincipalOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "用户账户"
        ],
        "summary": "更新当前用户资料",
        "operationId": "update_me_api_v1_users_me_patch",
        "parameters": [
          {
            "name": "pinjie_blog_web_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UserPrincipalOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "用户账户"
        ],
        "summary": "注销当前用户账户并移入回收站",
        "operationId": "delete_account_api_v1_users_me_delete",
        "parameters": [
          {
            "name": "pinjie_blog_web_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountDeleteIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ActionResult_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/avatar": {
      "put": {
        "tags": [
          "用户账户"
        ],
        "summary": "更新当前用户头像",
        "operationId": "update_avatar_api_v1_users_me_avatar_put",
        "parameters": [
          {
            "name": "pinjie_blog_web_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserAvatarUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UserPrincipalOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/password": {
      "post": {
        "tags": [
          "用户账户"
        ],
        "summary": "修改当前用户密码",
        "operationId": "change_password_api_v1_users_me_password_post",
        "parameters": [
          {
            "name": "pinjie_blog_web_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_RefreshSessionOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/sessions": {
      "get": {
        "tags": [
          "用户账户"
        ],
        "summary": "获取当前用户会话列表",
        "operationId": "list_sessions_api_v1_users_me_sessions_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "pinjie_blog_web_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_PageResult_SessionRead__"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/sessions/{session_id}": {
      "delete": {
        "tags": [
          "用户账户"
        ],
        "summary": "撤销当前用户的指定会话",
        "operationId": "revoke_session_api_v1_users_me_sessions__session_id__delete",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "pinjie_blog_web_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ActionResult_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/users/me/sessions/revoke-others": {
      "post": {
        "tags": [
          "用户账户"
        ],
        "summary": "撤销当前用户的其他会话",
        "operationId": "revoke_other_sessions_api_v1_users_me_sessions_revoke_others_post",
        "parameters": [
          {
            "name": "pinjie_blog_web_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Web Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ActionResult_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "后台管理"
        ],
        "summary": "获取用户列表",
        "operationId": "list_users_api_v1_admin_users_get",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "lifecycle",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "active",
                "inactive",
                "deleted"
              ],
              "type": "string",
              "default": "all",
              "title": "Lifecycle"
            }
          },
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_PageResult_AdminUserRead__"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "后台管理"
        ],
        "summary": "创建用户",
        "description": "由具备独立权限的管理员创建普通用户账户，不受公开注册开关影响且不创建登录会话。",
        "operationId": "create_user_api_v1_admin_users_post",
        "parameters": [
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminUserCreateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_AdminUserRead_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users/{user_id}": {
      "get": {
        "tags": [
          "后台管理"
        ],
        "summary": "获取用户详情",
        "operationId": "get_user_api_v1_admin_users__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UserPrincipalOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "后台管理"
        ],
        "summary": "更新用户资料",
        "operationId": "update_user_api_v1_admin_users__user_id__patch",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserUpdateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_UserPrincipalOut_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/users/{user_id}/sessions": {
      "get": {
        "tags": [
          "后台管理"
        ],
        "summary": "获取用户会话列表",
        "operationId": "list_user_sessions_api_v1_admin_users__user_id__sessions_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_PageResult_SessionRead__"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/users/{user_id}/sessions/{session_id}": {
      "delete": {
        "tags": [
          "后台管理"
        ],
        "summary": "撤销用户指定会话",
        "operationId": "revoke_user_session_api_v1_admin_users__user_id__sessions__session_id__delete",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "User Id"
            }
          },
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Session Id"
            }
          },
          {
            "name": "pinjie_blog_admin_access",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Pinjie Blog Admin Access"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "请求成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseModel_ActionResult_"
                }
              }
            }
          },
          "422": {
            "description": "请求参数校验失败",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountDeleteIn": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Current Password",
            "description": "当前密码，最多 64 个字符"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_password"
        ],
        "title": "AccountDeleteIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail",
            "description": "请求参数校验错误详情列表"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PasswordChangeIn": {
        "properties": {
          "current_password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Current Password",
            "description": "当前密码，最多 64 个字符"
          },
          "new_password": {
            "type": "string",
            "maxLength": 64,
            "minLength": 6,
            "title": "New Password",
            "description": "新密码，长度为 6 至 64 个字符"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "current_password",
          "new_password"
        ],
        "title": "PasswordChangeIn"
      },
...
      "ResponseModel_ActionResult_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/ActionResult",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[ActionResult]"
      },
...
      "ResponseModel_PageResult_SessionRead__": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/PageResult_SessionRead_",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[PageResult[SessionRead]]"
      },
...
      "ResponseModel_RefreshSessionOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/RefreshSessionOut",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[RefreshSessionOut]"
      },
...
      "ResponseModel_UserPrincipalOut_": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "稳定程序代码"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "面向调用方的中文结果消息"
          },
          "data": {
            "$ref": "#/components/schemas/UserPrincipalOut",
            "description": "响应业务数据"
          },
          "request_id": {
            "type": "string",
            "title": "Request Id",
            "description": "用于定位本次请求的唯一标识"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "code",
          "message",
          "data",
          "request_id"
        ],
        "title": "ResponseModel[UserPrincipalOut]"
      },
...
      "UserAvatarUpdateIn": {
        "properties": {
          "asset_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Asset Id",
            "description": "本人上传的头像资产唯一标识，传 null 表示移除"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserAvatarUpdateIn"
      },
...
      "UserUpdateIn": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name",
            "description": "展示名称"
          },
          "email": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 320
              },
              {
                "type": "null"
              }
            ],
            "title": "Email",
            "description": "电子邮箱地址"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "UserUpdateIn"
      },
...
    }
  },
...
}